structure changes - files moved/renamed
