+++ app/src/main/java/com/example/checklist/presenter/MainPresenterInterface.java
@@ -4,6 +4,6 @@ import com.example.checklist.model.Task;
 
 import java.util.List;
 
-public interface Presenter {
+public interface MainPresenterInterface {
     void displayTaskList(List<Task> taskList);
 }
