+++ app/src/main/java/com/example/checklist/presenter/RVPresenterInterface.java
@@ -1,9 +1,9 @@
-package com.example.checklist.views.sub;
+package com.example.checklist.presenter;
 
 import com.example.checklist.model.Task;
 
 import java.util.List;
 
-public interface RVPresenter {
+public interface RVPresenterInterface {
 //    void displayList(boolean updateList);
 }
