--- app/src/main/java/com/example/checklist/views/NewTaskFragment.java
@@ -1,35 +1,0 @@
-package com.example.checklist.views;
-
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.Button;
-import android.widget.EditText;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
-
-import com.example.checklist.R;
-
-public class NewTaskFragment extends Fragment {
-    private EditText taskTitleEt;
-    private Button addBtn;
-
-    @Nullable
-    @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.frag_layout, container, false);
-        addBtn = view.findViewById(R.id.add);
-        return view;
-    }
-
-    @Override
-    public void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        addBtn.setOnClickListener(v -> {
-
-        });
-    }
-}
